Move legacy logging code back into logging.go
