--- base/logging_legacy.go
@@ -1,342 +1,0 @@
-//  Copyright (c) 2012-2013 Couchbase, Inc.
-//  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
-//  except in compliance with the License. You may obtain a copy of the License at
-//    http://www.apache.org/licenses/LICENSE-2.0
-//  Unless required by applicable law or agreed to in writing, software distributed under the
-//  License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-//  either express or implied. See the License for the specific language governing permissions
-//  and limitations under the License.
-
-package base
-
-import (
-	"errors"
-	"fmt"
-	"math"
-	"net/http"
-	"net/url"
-	"runtime"
-	"strings"
-
-	"github.com/couchbase/clog"
-	"github.com/couchbase/goutils/logging"
-)
-
-var errMarshalNilLevel = errors.New("can't marshal a nil *Level to text")
-
-const (
-	// If true, all HTTP request/response bodies will be logged.
-	// Use this sparingly as it will probably dump sensitive information into the logs.
-	EnableLogHTTPBodies = false
-)
-
-type Level int32
-
-// DeferredLogFn is an anonymous function that can be executed at a later date to log something.
-type DeferredLogFn func()
-
-//By setting DebugLevel to -1, if LogLevel is not set in the logging config it
-//will default to the zero value for int32 (0) which will disable debug
-//logging, InfoLevel logging will be the default output.
-const (
-	// DebugLevel logs are typically voluminous, and are usually disabled in
-	// production.
-	DebugLevel Level = iota - 1
-	// InfoLevel is the default logging priority.
-	InfoLevel
-	// WarnLevel logs are more important than Info, but don't need individual
-	// human review.
-	WarnLevel
-	// ErrorLevel logs are high-priority. If an application is running smoothly,
-	// it shouldn't generate any error-level logs.
-	ErrorLevel
-	// PanicLevel logs a message, then panics.
-	PanicLevel
-	// FatalLevel logs a message, then calls os.Exit(1).
-	FatalLevel
-)
-
-// sgLevel returns a compatible internal SyncGateway Log Level for
-// the given logging config Level
-// The mapping is:
-//
-// DebugLevel	-1 --> 1
-// InfoLevel 	 0 --> 1
-// WarnLevel 	 1 --> 2
-// ErrorLevel 	 2 --> 2
-// PanicLevel 	 3 --> 3
-// FatalLevel 	 4 --> 3
-//
-// This can be mapped by addition of 2 to level value
-// and then a division by 2 and return the ceil of the result
-// to round to nearest int sgLevel value
-func (l Level) sgLevel() int {
-	return int(math.Ceil(float64(l+2) / float64(2)))
-}
-
-// cgLevel returns a compatible go-couchbase/golog Log Level for
-// the given logging config Level
-// The mapping is:
-//
-// DebugLevel	-1 --> 7
-// InfoLevel 	 0 --> 4
-// WarnLevel 	 1 --> 3
-// ErrorLevel 	 2 --> 2
-// PanicLevel 	 3 --> 1
-// FatalLevel 	 4 --> 0
-func (l Level) cgLevel() logging.Level {
-	switch l {
-	case DebugLevel:
-		return logging.DEBUG
-	case InfoLevel:
-		return logging.INFO
-	case WarnLevel:
-		return logging.WARN
-	case ErrorLevel:
-		return logging.ERROR
-	case PanicLevel:
-		return logging.SEVERE
-	case FatalLevel:
-		return logging.SEVERE
-	default:
-		return logging.NONE
-	}
-}
-
-// String returns a lower-case ASCII representation of the log level.
-func (l Level) String() string {
-	switch l {
-	case DebugLevel:
-		return "debug"
-	case InfoLevel:
-		return "info"
-	case WarnLevel:
-		return "warn"
-	case ErrorLevel:
-		return "error"
-	case PanicLevel:
-		return "panic"
-	case FatalLevel:
-		return "fatal"
-	default:
-		return fmt.Sprintf("Level(%d)", l)
-	}
-}
-
-// Implementaion of the go encoding.TextMarshaller interface for the Level type
-// This method will also be called by the JSON Marshaller
-//
-// MarshalText marshals the Level to text. Note that the text representation
-// drops the -Level suffix (see example).
-func (l *Level) MarshalText() ([]byte, error) {
-	if l == nil {
-		return nil, errMarshalNilLevel
-	}
-	return []byte(l.String()), nil
-}
-
-// Implementaion of the go encoding.TextUnmarshaller interface for the Level type
-// This method will also be called by the JSON Unmarshaller e.g. when loading
-// from logging configuration.
-//
-// UnmarshalText unmarshals text to a level. Like MarshalText, UnmarshalText
-// expects the text representation of a Level to drop the -Level suffix (see
-// example).
-//
-// In particular, this makes it easy to configure logging levels using YAML,
-// TOML, or JSON files.
-func (l *Level) UnmarshalText(text []byte) error {
-	switch string(text) {
-	case "debug":
-		*l = DebugLevel
-	case "info":
-		*l = InfoLevel
-	case "warn":
-		*l = WarnLevel
-	case "error":
-		*l = ErrorLevel
-	case "panic":
-		*l = PanicLevel
-	case "fatal":
-		*l = FatalLevel
-	default:
-		return fmt.Errorf("unrecognized level: %v", string(text))
-	}
-	return nil
-}
-
-type LogRotationConfig struct {
-	// MaxSize is the maximum size in megabytes of the log file before it gets
-	// rotated. It defaults to 100 megabytes.
-	MaxSize int `json:",omitempty"`
-
-	// MaxAge is the maximum number of days to retain old log files based on the
-	// timestamp encoded in their filename.  Note that a day is defined as 24
-	// hours and may not exactly correspond to calendar days due to daylight
-	// savings, leap seconds, etc. The default is not to remove old log files
-	// based on age.
-	MaxAge int `json:",omitempty"`
-
-	// MaxBackups is the maximum number of old log files to retain.  The default
-	// is to retain all old log files (though MaxAge may still cause them to get
-	// deleted.)
-	MaxBackups int `json:",omitempty"`
-
-	// LocalTime determines if the time used for formatting the timestamps in
-	// backup files is the computer's local time.  The default is to use UTC
-	// time.
-	LocalTime bool `json:",omitempty"`
-	// contains filtered or unexported fields
-}
-
-type LogAppenderConfig struct {
-	// Filename is the file to write logs to.  Backup log files will be retained
-	// in the same directory.  It uses <processname>-lumberjack.log in
-	// os.TempDir() if empty.
-	LogFilePath    *string            `json:",omitempty"`
-	LogKeys        []string           `json:",omitempty"` // Log keywords to enable
-	LogLevel       Level              `json:",omitempty"`
-	Rotation       *LogRotationConfig `json:",omitempty"`
-	RedactionLevel RedactionLevel     `json:",omitempty"`
-}
-
-// For transforming a new log level to the old type.
-func ToDeprecatedLogLevel(logLevel LogLevel) *Level {
-	var deprecatedLogLevel Level
-	switch logLevel {
-	case LevelDebug:
-		deprecatedLogLevel = DebugLevel
-	case LevelInfo:
-		deprecatedLogLevel = InfoLevel
-	case LevelWarn:
-		deprecatedLogLevel = WarnLevel
-	case LevelError:
-		deprecatedLogLevel = ErrorLevel
-	}
-	return &deprecatedLogLevel
-}
-
-// For transforming an old log level to the new type.
-func ToLogLevel(deprecatedLogLevel Level) *LogLevel {
-	var newLogLevel LogLevel
-	switch deprecatedLogLevel {
-	case DebugLevel:
-		newLogLevel.Set(LevelDebug)
-	case InfoLevel:
-		newLogLevel.Set(LevelInfo)
-	case WarnLevel:
-		newLogLevel.Set(LevelWarn)
-	case ErrorLevel:
-		newLogLevel.Set(LevelError)
-	}
-	return &newLogLevel
-}
-
-func EnableSgReplicateLogging() {
-	clog.EnableKey("Replicate")
-}
-
-func DisableSgReplicateLogging() {
-	clog.DisableKey("Replicate")
-}
-
-// GetLogKeys returns log keys in a map
-func GetLogKeys() map[string]bool {
-	consoleLogKeys := ConsoleLogKey().EnabledLogKeys()
-	logKeys := make(map[string]bool, len(consoleLogKeys))
-	for _, v := range consoleLogKeys {
-		logKeys[v] = true
-	}
-	return logKeys
-}
-
-// UpdateLogKeys updates the console's log keys from a map
-func UpdateLogKeys(keys map[string]bool, replace bool) {
-	if replace {
-		ConsoleLogKey().Set(KeyNone)
-	}
-
-	for k, v := range keys {
-		key := strings.Replace(k, "+", "", -1)
-		if v {
-			ConsoleLogKey().Enable(logKeyNamesInverse[key])
-		} else {
-			ConsoleLogKey().Disable(logKeyNamesInverse[key])
-		}
-	}
-
-	Infof(KeyAll, "Setting log keys to: %v", ConsoleLogKey().EnabledLogKeys())
-}
-
-// Returns a string identifying a function on the call stack.
-// Use depth=1 for the caller of the function that calls GetCallersName, etc.
-func GetCallersName(depth int, includeLine bool) string {
-	pc, file, line, ok := runtime.Caller(depth + 1)
-	if !ok {
-		return "???"
-	}
-
-	fnname := ""
-	if fn := runtime.FuncForPC(pc); fn != nil {
-		fnname = lastComponent(fn.Name())
-	}
-
-	if !includeLine {
-		return fnname
-	}
-
-	return fmt.Sprintf("%s() at %s:%d", fnname, lastComponent(file), line)
-}
-
-// Partial interface for the SGLogger
-type SGLogger interface {
-	Logf(logLevel LogLevel, logKey LogKey, format string, args ...interface{})
-}
-
-func lastComponent(path string) string {
-	if index := strings.LastIndex(path, "/"); index >= 0 {
-		path = path[index+1:]
-	} else if index = strings.LastIndex(path, "\\"); index >= 0 {
-		path = path[index+1:]
-	}
-	return path
-}
-
-// This provides an io.Writer interface around the base.Infof API
-type LoggerWriter struct {
-	LogKey       LogKey        // The log key to log to, eg, KeyHTTP
-	SerialNumber uint64        // The request ID
-	Request      *http.Request // The request
-	QueryValues  url.Values    // A cached copy of the URL query values
-}
-
-// Write() method to satisfy the io.Writer interface
-func (lw *LoggerWriter) Write(p []byte) (n int, err error) {
-	Infof(lw.LogKey, " #%03d: %s %s %s", lw.SerialNumber, lw.Request.Method, SanitizeRequestURL(lw.Request, &lw.QueryValues), string(p))
-	return len(p), nil
-}
-
-// Create a new LoggerWriter
-func NewLoggerWriter(logKey LogKey, serialNumber uint64, req *http.Request, queryValues url.Values) *LoggerWriter {
-	return &LoggerWriter{
-		LogKey:       logKey,
-		SerialNumber: serialNumber,
-		Request:      req,
-		QueryValues:  queryValues,
-	}
-}
-
-// Prepend a context ID to each blip logging message.  The contextID uniquely identifies the blip context, and
-// is useful for grouping the blip connections in the log output.
-func PrependContextID(contextID, format string, params ...interface{}) (newFormat string, newParams []interface{}) {
-
-	// Add a new format placeholder for the context ID, which should appear at the beginning of the logs
-	formatWithContextID := `[%s] ` + format
-
-	params = append(params, 0)
-	copy(params[1:], params[0:])
-	params[0] = contextID
-
-	return formatWithContextID, params
-
-}
